backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
paddleocr
python-dotenv
requests
numpy